MIPS: pci-ar724x: add QCA9550 reset sequence
The QCA9550 family of SoCs have a slightly different reset
sequence compared to older chips.

Normally the bootloader performs this sequence, however
some bootloader implementation expect the operating system
to clear the reset.

Also get the resets from OF to support handling of the second
PCIe root-complex on the QCA9558.

Signed-off-by: David Bauer <mail@david-bauer.net>